pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "ali-cli"
version = "0.1.0"
description = "CLI tool for Alibaba.com buyer portal automation (messenger, RFQs, quotes)"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
authors = [
    { name = "Derrek Wiedeman" }
]
keywords = ["alibaba", "rfq", "playwright", "browser-automation", "sourcing"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "playwright>=1.40",
    "click>=8.0",
    "rich>=13.0",
    "requests>=2.28",
    "google-auth>=2.0",
    "google-auth-oauthlib>=1.0",
    "google-api-python-client>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.1",
]

[project.scripts]
ali = "ali_cli.cli:main"

[project.urls]
Homepage = "https://github.com/gtg708q/ali-cli"
Issues = "https://github.com/gtg708q/ali-cli/issues"

[tool.setuptools.packages.find]
include = ["ali_cli*"]
exclude = ["archive*", "tests*", "scripts*", "docs*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP"]
ignore = ["E501"]  # line-length handled by formatter
